fix mistype

+++ src/main/java/org/guzz/orm/type/FloatObjectSQLDataType.java
@@ -73,7 +73,7 @@ public class FloatObjectSQLDataType implements SQLDataType {
 	}
 	
 	public Class getDataType(){
-		return Long.class ;
+		return Float.class ;
 	}
 
 	public Object getFromString(String value) {
